pyproject.toml
[tool.poetry]
name = "harley-bot"
version = "0.1.0"
description = "A multipurpose discord bot."
authors = ["Imvaskel <contact@vaskel.gay>"]
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = {git = "https://github.com/rapptz/discord.py"}
asyncpg = "0.21.0"
discord-ext-ipc = "2.0.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus"}
discord-ext-menus-views = {git = "https://github.com/oliver-ni/discord-ext-menus-views"}
jishaku = {git = "https://github.com/Gorialis/jishaku"}
"mystbin.py" = "2.1.0"
Markdown = "3.3.3"
asyncdagpi = "^4.0.0"
coloredlogs = "^15.0.1"
psutil = "^5.8.0"
polaroid = "^0.3.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
